mypage간단 작성 컨트롤러 수정

+++ src/main/java/com/sds/foodfit/domain/MemberDetail.java
@@ -7,8 +7,8 @@ public class MemberDetail {
 	private int memberDetailIdx;
 	private String gender;
 	private int age;
-	private long height;
-	private long weight;	
+	private int height;
+	private int weight;	
 	private Member member; // 회원정보 객체 가짐
 
 	private DislikedFood dislikedFood; // 비선호 음식 객체 가짐
